requirements.txt
# 核心框架
crewai>=0.28.0
langchain>=0.1.0
langchain-openai>=0.1.0

# 图像处理
opencv-python>=4.8.0
Pillow>=10.0.0
numpy>=1.24.0

# 深度学习推理
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0

# API调用
requests>=2.31.0
openai>=1.10.0

# 工具库
pydantic>=2.0.0
python-dotenv>=1.0.0
loguru>=0.7.0

# 可选：如果使用本地模型
# ultralytics>=8.0.0  # YOLO系列模型
# segment-anything>=1.0  # SAM模型
